requirements.txt
# local package
-e .

# external requirements
# click
# alabaster==0.7.12
# alembic==1.4.1
# appdirs==1.4.4
# APScheduler==3.7.0
# atpublic==2.1.3
# attrs==20.3.0
autopep8==1.5.5
# awscli==1.19.27
# Babel==2.9.0
# botocore==1.20.27
# cached-property==1.5.2
# cachetools==4.2.1
certifi==2020.12.5
# cffi==1.14.5
# chardet==4.0.0
# cloudpickle==1.6.0
# colorama==0.4.3
# colorhash==1.0.3
# commonmark==0.9.1
# configobj==5.0.6
# configparser==5.0.2
# coverage==5.5
# cryptography==3.4.6
# cycler==0.10.0
# databricks-cli==0.14.2
# decorator==4.4.2
# dictdiffer==0.8.1
# diskcache==5.2.1
# distlib==0.3.1
# distro==1.5.0
# docker==4.4.4
# docutils==0.15.2
# dpath==2.0.1
# dulwich==0.20.20
# dvc==2.0.5
# entrypoints==0.3
filelock==3.0.12
flake8==3.9.0
Flask==1.1.2
Flask-Cors==3.0.10
# Flask-MonitoringDashboard==3.1.0
flatten-dict==0.3.0
flufl.lock==3.2
fsspec==0.8.7
ftfy==5.9
funcy==1.15
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
# google-api-core==1.26.1
# google-api-python-client==2.0.2
# google-auth==1.27.1
# google-auth-httplib2==0.1.0
# googleapis-common-protos==1.53.0
# grandalf==0.6
gunicorn==20.0.4
httplib2==0.19.0
# idna==2.10
# imagesize==1.2.0
# importlib-metadata==3.7.3
# iniconfig==1.1.1
# itsdangerous==1.1.0
# Jinja2==2.11.1
# jmespath==0.10.0
joblib==1.0.1
jsonpath-ng==1.5.2
# kiwisolver==1.3.1
# kneed==0.7.0
# livereload==2.6.3
# lunr==0.5.8
# mailchecker==4.0.3
# Mako==1.1.4
# Markdown==3.3.4
# MarkupSafe==1.1.1
# matplotlib==3.3.4
# mccabe==0.6.1
# mkdocs==1.1.2
# mkdocs-material==7.0.6
# mkdocs-material-extensions==1.0.1
# mlflow==1.14.1
# nanotime==0.5.2
# networkx==2.5
# nltk==3.5
# numpy==1.20.1
# oauth2client==4.1.3
# packaging==20.9
pandas==1.2.3
# pathlib2==2.3.5
# pathspec==0.8.1
# phonenumbers==8.12.19
# Pillow==8.1.2
# pluggy==0.13.1
# ply==3.11
# prometheus-client==0.9.0
# prometheus-flask-exporter==0.18.1
# protobuf==3.15.6
# psutil==5.8.0
# py==1.10.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycodestyle==2.7.0
# pycparser==2.20
# pydantic==1.8.1
# pydot==1.4.2
# PyDrive2==1.8.1
# pyflakes==2.3.0
# pygit2==1.5.0
# Pygments==2.8.1
# pygtrie==2.3.2
# pymdown-extensions==8.1.1
pyOpenSSL==20.0.1
# pyparsing==2.4.7
# pytest==6.2.2
# python-benedict==0.23.2
# python-dateutil==2.8.1
# python-dotenv==0.15.0
# python-editor==1.0.4
# python-fsutil==0.4.0
# python-slugify==4.0.1
# pytz==2021.1
PyYAML==5.4.1
# querystring-parser==1.2.4
# regex==2020.11.13
# requests==2.25.1
# rich==9.13.0
# rsa==4.7.2
# ruamel.yaml==0.16.13
# ruamel.yaml.clib==0.2.2
# s3transfer==0.3.4
scikit-learn==0.24.1
# scipy==1.6.1
# shortuuid==1.0.1
# shtab==1.3.5
# six==1.15.0
# smmap==3.0.5
# snowballstemmer==2.1.0
# Sphinx==3.5.2
# sphinxcontrib-applehelp==1.0.2
# sphinxcontrib-devhelp==1.0.2
# sphinxcontrib-htmlhelp==1.0.3
# sphinxcontrib-jsmath==1.0.1
# sphinxcontrib-qthelp==1.0.3
# sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.23
sqlparse==0.4.1
# tabulate==0.8.9
# text-unidecode==1.3
# threadpoolctl==2.1.0
# toml==0.10.2
# tornado==6.1
# tox==3.23.0
# tqdm==4.59.0
# typing-extensions==3.7.4.3
# tzlocal==2.1
# uritemplate==3.0.1
# urllib3==1.26.3
# virtualenv==20.4.2
# voluptuous==0.12.1
# wcwidth==0.2.5
# websocket-client==0.58.0
# Werkzeug==1.0.1
# xmltodict==0.12.0
# zc.lockfile==2.0
# zipp==3.4.1